rename infobox "callout", swap in place for scss module
