+++ airbyte-webapp/src/components/ui/Alert/Alert.tsx
@@ -8,7 +8,7 @@ interface AlertProps {
   className?: string;
 }
 
-export const Alert: React.FC<PropsWithChildren<AlertProps>> = ({ variant, children, className }) => {
+export const Alert: React.FC<PropsWithChildren<AlertProps>> = ({ variant = "blue", children, className }) => {
   const containerStyles = classNames(styles.container, className, {
     [styles.yellow]: variant === "yellow",
     [styles.blue]: variant === "blue",
