+++ airbyte-webapp/src/components/ui/Callout/index.ts
@@ -1,0 +1,1 @@
+export { Callout as InfoBox } from "./Callout";
