--- airbyte-webapp/src/components/ui/InfoBox/index.ts
@@ -1,1 +1,0 @@
-export { InfoBox } from "./InfoBox";
